--- src/search/components/XButton.tsx
@@ -1,22 +1,0 @@
-import { StyledXButton } from './SearchBar/SearchBar.style';
-
-const XButton = ({ initiateValue }: { initiateValue: () => void }) => {
-  return (
-    <StyledXButton onClick={() => initiateValue()}>
-      <svg
-        width="20"
-        height="20"
-        viewBox="0 0 20 20"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M10 0C8.02219 0 6.08879 0.58649 4.4443 1.6853C2.79981 2.78412 1.51809 4.3459 0.761209 6.17317C0.00433284 8.00043 -0.193701 10.0111 0.192152 11.9509C0.578004 13.8907 1.53041 15.6725 2.92894 17.0711C4.32746 18.4696 6.10929 19.422 8.0491 19.8079C9.98891 20.1937 11.9996 19.9957 13.8268 19.2388C15.6541 18.4819 17.2159 17.2002 18.3147 15.5557C19.4135 13.9112 20 11.9778 20 10C20 8.68678 19.7413 7.38642 19.2388 6.17317C18.7363 4.95991 17.9997 3.85752 17.0711 2.92893C16.1425 2.00035 15.0401 1.26375 13.8268 0.761205C12.6136 0.258658 11.3132 0 10 0ZM14.53 13.47C14.6705 13.6106 14.7493 13.8012 14.7493 14C14.7493 14.1988 14.6705 14.3894 14.53 14.53C14.3894 14.6705 14.1988 14.7493 14 14.7493C13.8013 14.7493 13.6106 14.6705 13.47 14.53L10 11.06L6.53 14.53C6.38938 14.6705 6.19875 14.7493 6 14.7493C5.80125 14.7493 5.61063 14.6705 5.47 14.53C5.32955 14.3894 5.25066 14.1988 5.25066 14C5.25066 13.8012 5.32955 13.6106 5.47 13.47L8.94 10L5.47 6.53C5.33752 6.38783 5.2654 6.19978 5.26883 6.00548C5.27226 5.81118 5.35097 5.62579 5.48838 5.48838C5.6258 5.35097 5.81118 5.27225 6.00548 5.26883C6.19978 5.2654 6.38783 5.33752 6.53 5.47L10 8.94L13.47 5.47C13.6122 5.33752 13.8002 5.2654 13.9945 5.26883C14.1888 5.27225 14.3742 5.35097 14.5116 5.48838C14.649 5.62579 14.7278 5.81118 14.7312 6.00548C14.7346 6.19978 14.6625 6.38783 14.53 6.53L11.06 10L14.53 13.47Z"
-          fill="#B5B9BD"
-        />
-      </svg>
-    </StyledXButton>
-  );
-};
-
-export default XButton;
